Also move plugin selection logic from display.ops
